+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and from version 0.9.0 onwards this project adheres to [Semantic Versioning](htt
 
 ## [Unreleased]
 
+### Added
+
+- Added export of any `wxListCtrl` directly to PDF via `wxPdfDocument::AddList()` or `wxPdfDC::DrawList()`
+- Added VC++ 2026 build support
+
 ## [1.4.0] - 2026-05-20
 
 ### Changed